+++ source/modules/thmc/two_phase_flow/HeterogeneityAndRateAwareSaturationFunctions.cpp
@@ -1183,7 +1183,7 @@ double HeterogeneityAndRateAwareSaturationFunctions<dim,USER>::dpcds_at_VG( Elem
  
  
 template<uint32_t dim, template<uint32_t> class USER>
-double HeterogeneityAndRateAwareSaturationFunctions<dim,USER>::krw_parallel( Element<dim>* const e, size_t direction ) const
+double HeterogeneityAndRateAwareSaturationFunctions<dim,USER>::krw_parallel( Element<dim>* const e, int direction ) const
 {
     ErrorHandler&  csmp_error( ErrorHandler::Instance() );
